Fixing the ugly names for, well, everything
We changed the names of the apks and distZips to be human readable and
meaningful. We also changed the names for the execution scripts in the
distZips to be nice. We also got rid of the extra directory level in
the distZips.

I'm embarassed at just how much time this all took to get working. I
really have to figure out how to get along with Gradle a lot better.